==> 1_Convert_to_CSV.ipynb <==
import pandas as pd
import json

# Read the entire JSON file and loading all JSON objects
with open("./train.json", 'r') as json_file:
    json_data = json_file.read()

# Split the file content into individual JSON objects
json_objects = json_data.strip().split('\n')

# Convert JSON objects to DataFrames and store them in a list
dataframes = []
for json_obj in json_objects:
    try:
        json_dict = json.loads(json_obj)
        df = pd.DataFrame([json_dict])
        dataframes.append(df)
    except json.JSONDecodeError:
        print(f"Skipping invalid JSON: {json_obj}")

# Concatenate all DataFrames into one
final_dataframe = pd.concat(dataframes, ignore_index=True)

# Save the combined DataFrame to a CSV file
final_dataframe.to_csv('output.csv', index=False)

print("Combined DataFrame saved as output.csv")

import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

import pandas as pd
df=pd.read_csv("output.csv")

df.shape

df.head()

df=df.drop(["choice_type","subject_name","topic_name","exp"],axis=1)
df.head()

df.isnull().sum()

df.shape

df=df.dropna(subset="opa")
df=df.dropna(subset="opb")
df=df.dropna(subset="opc")
df=df.dropna(subset="opd")

print(df.shape)
df.isnull().sum()

df.head()

random_diff=np.random.randint(0,3,size=len(df))
df["Difficulty"]=random_diff
df.head()

df.to_csv('Questions.csv', index=False)
