==> 2_Question_Level.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import itertools

# Generate all unique combinations of three numbers that sum to 10
combinations = [combo for combo in itertools.product(range(11), repeat=3) if sum(combo) == 10]

# Define the number of rows
num_rows = 100

# Randomly sample from the combinations to create the DataFrame
np.random.shuffle(combinations)
data = combinations[:num_rows]

df = pd.DataFrame(data, columns=['n_beginner', 'n_intermediate', 'n_advance'])

print(df)


for i in range(0,11):
    df2=df.copy()
    df=pd.concat([df,df2])
print(df)

import random

# Generate a random number between 0 and 1
for i in ["accuracy_beginner","accuracy_intermediate","accuracy_advance"]:
    random_number = np.random.uniform(0,1,135168)
    df[i]=random_number

df

df.info()

df.loc[df['n_beginner'] == 0, 'accuracy_beginner'] = 1.0
df.loc[df['n_intermediate'] == 0, 'accuracy_intermediate'] = 1.0
df.loc[df['n_advance'] == 0, 'accuracy_advance'] = 1.0


df

df.to_csv("Question_level.csv")
