==> 3_Final_dataset.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")

df=pd.read_csv("C:/Users/PC-ACER/DataScience/DataScienceProject/Question_level.csv")
df

df=df.drop("Unnamed: 0",axis=1)
df

#decide prev level
def decide_prev_level(current_level):
    index=df.columns.get_loc(current_level)
    previous_level_index=0
    if index==0:
        previous_level_index=index
    else:
        previous_level_index=index-1
    return previous_level_index

#decide next level
def decide__next_level(current_level):
    index=df.columns.get_loc(current_level)
    next_level_index=0
    if index==2:
        next_level_index=index
    else:
        next_level_index=index+1
    return next_level_index
    

#update no. of question in level
def update_n(df,beg_n,beg_percent,int_n,int_percent,adv_n,adv_percent,new_beginner,new_intermediate,new_advance):
    
    #####################################################################
    # for beginner : level switch - upgrade since no prev : for acc_beg>0
                     # level switch - degrade since no prev : for acc_beg<0
    #######################################################################
    beg_new=(-1) * round(beg_percent*beg_n)
    diff=round(beg_percent*beg_n)
    n_int_beg=diff
    
    
    #################################################################
    # for intermediate : level switch - upgrade or INC : depends on weight
    ##########################################################################

    int_new=int_n

    if int_percent > 0:
        #level switch - upgrade
        int_new_LSU=(-1) * round(int_percent*int_n)
        diff=round(int_percent*int_n)
        n_adv_int=diff
    
        #INC
        int_new_INC=round(int_percent*int_n)
        diff=round(int_percent*int_n)
        if(diff>beg_n):
            while diff<=beg_n:
                int_new_INC-=1
                diff-=1
        n_beg_int=(-1) * diff
    
        #deciding weightage : Beginner : 28% , Intermediate : 34% , Advance : 38%
        weight_LSU= (int_new_LSU * 0.34) + (n_adv_int * 0.38)
        weight_INC= (int_new_INC * 0.34) + (n_beg_int * 0.28)
    
        if(weight_LSU > weight_INC):
            int_new=int_new_LSU
            n_adv_int=n_adv_int
            n_beg_int=0
        else:
            int_new=int_new_INC
            n_adv_int=0
            n_beg_int=n_beg_int
    else:
        #DEC
        int_new=round(int_percent*int_n)
        diff=round(int_percent*int_n)
        n_beg_int=(-1) * diff
        n_adv_int=0
    
    #############################################################################
    # for advance : INC since no next
    #############################################################################
    if adv_percent > 0:
        #INC
        adv_new=round(adv_percent*adv_n)
        diff=round(adv_percent*adv_n)
        if(diff>int_n):
            while diff<=int_n:
                adv_new-=1
                diff-=1
        n_int_adv=(-1) * diff
    else:
        #DEC
        adv_new=round(adv_percent*adv_n)
        diff=round(adv_percent*adv_n)
        n_int_adv=(-1) * diff

    ###########################################################################
    # Updating values
    ###########################################################################

    # print(beg_n+beg_new + n_beg_int)
    # print(int_n+int_new + n_int_beg + n_int_adv)
    # print(adv_n+adv_new + n_adv_int)
    new_beginner.append(beg_n+beg_new + n_beg_int)
    new_intermediate.append(int_n+int_new + n_int_beg + n_int_adv)
    new_advance.append(adv_n+adv_new + n_adv_int)
      
                   
    
# adding column of percentage increase/decrease in each level
def ranges(accuracy,n):
    updated=0.0

    ########### level decrease ##########
    if accuracy>=0.00 and accuracy<0.10:
        updated= -0.72
    elif accuracy>=0.10 and accuracy<0.20:
        updated= -0.66
    elif accuracy>=0.20 and accuracy<0.30:
        updated= -0.60
    elif accuracy>=0.30 and accuracy<0.40:
        updated= -0.54
    elif accuracy>=0.40 and accuracy<0.45:
        updated= -0.48
    elif accuracy>=0.45 and accuracy<0.50:
        updated= -0.44
    elif accuracy>=0.50 and accuracy<0.55:
        updated= -0.40
    elif accuracy>=0.55 and accuracy<0.60:
        updated= -0.36

    ######### level increase ###############
    elif accuracy>=0.60 and accuracy<0.70:
        updated=0.35
    elif accuracy>=0.70 and accuracy<0.80:
        updated=0.44
    elif accuracy>=0.80 and accuracy<0.85:
        updated=0.51
    elif accuracy>=0.85 and accuracy<0.90:
        updated=0.58
    elif accuracy>=0.90 and accuracy<0.95:
        updated=0.65
    else:
        updated=0.75
    return updated

new_beginner=[]
new_intermediate=[]
new_advance=[]
for i in range(0,len(df)):
    new_beginner.append(ranges(df["accuracy_beginner"][i],df["n_beginner"][i]))
    new_intermediate.append(ranges(df["accuracy_intermediate"][i],df["n_intermediate"][i]))
    new_advance.append(ranges(df["accuracy_advance"][i],df["n_advance"][i]))
    
df["%_beginner"]=new_beginner
df["%_intermediate"]=new_intermediate
df["%_advance"]=new_advance
df

new_beginner=[]
new_intermediate=[]
new_advance=[]
for i in range(0,len(df)):
    update_n(df,df["n_beginner"][i],df["%_beginner"][i],df["n_intermediate"][i],df["%_intermediate"][i],df["n_advance"][i],df["%_advance"][i],new_beginner,new_intermediate,new_advance)
df["new_beginner"]=new_beginner
df["new_intermediate"]=new_intermediate
df["new_advance"]=new_advance

df

tempdf=df.copy()

tempdf = tempdf[tempdf['new_beginner'] >= 0]
tempdf = tempdf[tempdf['new_intermediate'] >= 0]
tempdf = tempdf[tempdf['new_advance'] >= 0]
tempdf


tempdf['sum_cols'] = tempdf['new_beginner'] + tempdf['new_intermediate'] + tempdf['new_advance']
tempdf

tempdf = tempdf[tempdf['sum_cols'] == 10]
tempdf

tempdf=tempdf.drop("sum_cols",axis=1)
tempdf

tempdf.to_csv("final_dataset.csv",index=False)
